Added a makefile for ease of compiling and running the package.
Renamed CsvParser to IsolateFileParser to be more descriptive.

Fixed bugs/errors in IsolateFileParser, ParserDriver, IsolateSamle,
SampleMethod.

These classes now are correctly packaged and have valid references to
each other such that the data given to me by Emily can be correctly
parsed and interpreted into a Map<Integer, List<IsolateSample>>
representing a mapping from days to a list of samples for that day.