Made several changes trying to ensure clustering clusters appropriately.
I changed corrDistance in Cluster.java to use correlation (100-based not
0-based) for distances.

I attempted to remove the concept of sampling methods being a numerical
value representing its moment in time in favor of splitting the groups
into separate lists, clustering in those lists, and then clustering the
union of F and I (fecal and immediate) and then clustering the union of
F_I and L (fecal+immediate and later). In this way I ensure that the
ordering is done correctly and I believe I avoid the problem of possibly
only clustering them once instead of continually clustering.

I want to eventually switch it back for the interest of conciseness and
performance, however I will do that when I have more time. As it is I
want to get good results so that I can submit to a conference