Modified how squishy isolates get integrated with established strongly connected clusters by adding them to the list of clusters as single isolate clusters. Then I call clusterGroup which goes through the list of clusters and clusters everything.
This also fixes fragmented clusters based on strongly connected isolates to be
more unified clusters when using squishy correlations. This increases average
correlation drastically which is how clusters are able to become larger.

In order to address the aforementioned problem, I added a method to
Cluster.java: setSimilarityMatrix. By populating the weakly connected matrix
with all of the correlations from the strongly connected matrix (change made in
IsolateFileParser.java), it is possible to use setSimilarityMatrix on all
strongly connected clusters to expand the correlation mappings for isolates and
to appropriately cluster squishy isolates as well as combine clusters.

Also, a method for generating highcharts-compatible 'series' output is in
progress in Cluster.java