committing newest version of jar file
This commit fixes the duplicate isolate issue that was appearing in the
output of the program. Additionally, an intermediate
IsolateSimilarityMatrix is used until a correlation between two isolates
consists of both the 16-23 and 23-5 regions in which case it is moved to
the appropriate isolateNetwork (STRONG or WEAK).

some things to note. IsolateSimilarityMatrix's contain
IsolateCorrelation's which contain a correlation for both regions. When
a correlation value is retrieved from the isolate similarity matrix it
does the appropriate transformation (> alpha, < beta) but doesn't change
the correlations value, it just transforms the value returned. This may
or may not cause problems so I am noting it here to investigate further.