Updating this program before I accidentally blow away awesome changes.
Modified thresholds so that lower and upper thresholds for a region are
set and appropriately used. Global threshold is being moved out of the
program and will somehow be re-implemented as a post-processing
visualization of the result dendogram.

Additionally there are 2 varying methods (both implemented in
Cluster.java as isSimilar() and isDifferent()) for determining what is
the clustering rule (since we can't allow all isolates/clusters in each
phase to fully cluster or we will never get variability in our
clusters). This needs to be investigated further and tested though.

Also IsolateCorrelation was not validly using region thresholds, it has
been modified to do so.