Got global thresholding to work with basic functionality.
The user is now able to apply a threshold to a dendogram file. This
implementation may be okay but there needs to be better feedback as far
as what has changed so that they don't have to open the file constantly
to see if the applied threshold is better or worse than another.

Changes detailed:
overloaded IsolateOutputWriter that prints temporalCharts. The new one
is specifically for the dendogram thresholding. This is because I made
it stupid hard to make a cluster and so this was a hacked solution.

DendogramLeaf was inconsistent in it's printing of xml. I changed data
attributes to be isolate attributes for consistency and maintainability.

DendogramParser was invalidly applying the threshold (It seems that when
you use getElementsByTagName it applies to allllllll sub nodes of that
tag name. I had to change this to getChildNodes()) and so this was
fixed.

ThresholdDialog was modified to correctly close itself after OK is
pressed and also supplies a more valid output file name, though it is
not specifiable yet.

Simply removed comments from ClusterSingleListener